+++ test/foundry/Raffle.fulfillRandomWords.t.sol
@@ -35,9 +35,9 @@ contract Raffle_FulfillRandomWords_Test is TestHelpers {
         vm.startPrank(user1);
         looksRareRaffle.createRaffle({
             cutoffTime: uint40(block.timestamp + 86_400),
-            minimumEntries: uint64(107),
-            maximumEntries: 512,
-            maximumEntriesPerParticipant: uint64(100),
+            minimumEntries: uint80(107),
+            maximumEntries: uint80(512),
+            maximumEntriesPerParticipant: uint80(100),
             prizesTotalValue: 1 ether,
             minimumProfitBp: uint16(500),
             feeTokenAddress: address(0),
